The maze and player each move properly, with a couple of minor things that need to be fixed.
Something worth mentioning, there could be problems later on with the
swipe gestures and how they are handled.  It appears that more than one
action can run simultaneously if the screen is swiped at the right time,
which would not work well with the logic for determining whether the
player or maze should move.