Added isPlayer property to MMCharacter, that way we can use that instead of the isKindOfClass method when checking to see if the character is a player.
Rearranged some of the logic for setting up an Enemy.  The level now
first only adds the enemies to the game layer, and then the positions
are set from the game layer.

Got the enemy moving stuff to work, double tapping a tile now causes the
enemy to path to it, and whether or not the main player is moving will
not affect this behavior, even if you can't see the enemy on the screen.

Also added more temporary images for the enemy -- the screen on their
helmets is red instead of green.