Lots of changes, right now there is only one enemy, the player moves super fast so it's easier to test.
The biggest changes from this commit is that each character has a new
property called a traveler key, which is used while the path finder
class is setting attributes on tiles during the A* algorithm.  Each
player needs to reference it's own set of tile attributes because tiles
cannot be shared for path finding between characters, other than their
address in memory.  For example, one enemy calculating a path will set
each tile's parent property to use when backtracing to the start, but
the parent tiles for that enemies path will most likely be completley
different than the parent tiles of another enemies path.  Since they're
using the same address in memory for each tile, then each tile needs to
contain a dictionary of tile attributes, where each tile attributes
dictionary will be associated with a traveler key.